requirements.txt
astroid==2.0.1
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.4
django-filter==2.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.4
PyJWT==1.6.4
pylint==2.0.1
python-dotenv==0.10.1
pytz==2018.3
requests==2.21.0
simplejson==3.13.2
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==3.3.1
wrapt==1.10.11
